Added learn-awesome/flashcard under Miscellaneous components
This set of components let the user practice through a set of flashcards. If you did not remember a card, it gets pushed at the end of the deck. Every card maintains its level in localStorage, and the increment/decrement logic implements Leitner-box like algorithm for spaced-repetition. The component supports library build, and is reusable by any webpage via just a `<script>` tag.